requirements.txt

Flask
tensorflow==2.20.0
Pillow
numpy
werkzeug